Routine commit
Netbox_route53.py tested with class and objects. Everything runs
smoothly

Added functionality to purge records that are in R53 and not in NB

Major script overhauls. Changed the way functions operated to
increase performance and reliability. Changed several variables
to be global and set all the self functions correctly.

IntegrateRecords.py

Work in progress